+++ lib/Templates/bottom_nav_bar.dart
